add more coverage

+++ app/Livewire/Components/Forms/Album/ShareWith.php
@@ -146,6 +146,9 @@ class ShareWith extends Component
 
 	public function delete(int $id): void
 	{
+		$perm = AccessPermission::with('album')->findOrFail($id);
+		Gate::authorize(AlbumPolicy::CAN_SHARE_WITH_USERS, [AbstractAlbum::class, $perm->album]);
+
 		AccessPermission::query()->where('id', '=', $id)->delete();
 		$this->resetData();
 	}
